requirements.txt
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tqdm==4.66.1
albumentations==1.3.1